+++ src/main/java/com/yuuko/core/events/controllers/GenericTextChannelController.java
@@ -1,8 +1,6 @@
 package com.yuuko.core.events.controllers;
 
 import com.yuuko.core.database.function.BindFunctions;
-import com.yuuko.core.database.function.DatabaseFunctions;
-import com.yuuko.core.database.function.GuildFunctions;
 import net.dv8tion.jda.api.events.channel.text.GenericTextChannelEvent;
 import net.dv8tion.jda.api.events.channel.text.TextChannelDeleteEvent;
 
@@ -15,23 +13,7 @@ public class GenericTextChannelController {
     }
 
     private void textChannelDeleteEvent(TextChannelDeleteEvent e) {
-        String channel = e.getChannel().getId();
-
-        BindFunctions.cleanupBinds(channel);
-
-        if(channel.equals(GuildFunctions.getGuildSetting("starboard", channel))) {
-            DatabaseFunctions.cleanupSettings("starboard", e.getGuild().getId());
-            return;
-        }
-
-        if(channel.equals(GuildFunctions.getGuildSetting("comlog", channel))) {
-            DatabaseFunctions.cleanupSettings("comlog", e.getGuild().getId());
-            return;
-        }
-
-        if(channel.equals(GuildFunctions.getGuildSetting("modlog", channel))) {
-            DatabaseFunctions.cleanupSettings("modlog", e.getGuild().getId());
-        }
+        BindFunctions.cleanupReferences(e.getChannel().getId());
     }
 }